requirements.txt
# Core framework
fastapi==0.116.1
uvicorn==0.35.0
starlette==0.47.3

# Database
sqlalchemy==2.0.36

# Pydantic (for schemas)
pydantic==2.11.7
pydantic-core==2.33.2

# Auth & security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.20

# Google / AI
google-generativeai==0.8.3
google-ai-generativelanguage==0.6.15
google-auth==2.40.3
google-api-core==2.25.1

# Utilities
requests==2.32.5
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
urllib3==2.5.0

google-cloud-speech==2.27.0
opencv-python
Pillow
google-cloud-aiplatform==1.71.1
